+++ tests/neuralnet/test_synapse_to_dict.py
@@ -1,4 +1,4 @@
-from decimal import Decimal
+
 import unittest
 
 from src.neuralnet.sigmoid_node import SigmoidNode
@@ -10,10 +10,10 @@ class TestSynapseToDict(unittest.TestCase):
     def test_synapse_to_dict(self):
         input_node = SigmoidNode()
         output_node = SigmoidNode()
-        weight = Decimal(0.5)
+        weight = float(0.5)
         synapse = Synapse(input_node, output_node, weight)
 
         synapse_dict = SynapseToDict().to_dict(synapse)
 
         self.assertEqual(synapse_dict['id'], synapse.id)
-        self.assertEqual(Decimal(synapse_dict.get('weight')), weight)
+        self.assertEqual(float(synapse_dict.get('weight')), weight)
