+++ tests/pipeline/test_competition_epoch.py
@@ -1,4 +1,4 @@
-from decimal import Decimal
+
 from typing import Union
 import unittest
 from unittest.mock import Mock
@@ -28,12 +28,12 @@ class TestCompetitionEpoch(unittest.TestCase):
         network_1 = Network([1, 1])
         network_2 = Network([1, 1])
         
-        network_1.synapse_layers[0][0].weight = Decimal(0)
-        network_2.synapse_layers[0][0].weight = Decimal(0)
+        network_1.synapse_layers[0][0].weight = float(0)
+        network_2.synapse_layers[0][0].weight = float(0)
         
         population = [network_1, network_2]
         
-        competition = MockCompetition(override_weights=Decimal(1))
+        competition = MockCompetition(override_weights=float(1))
         
         epoch = CompetitionEpoch(competition)
 
@@ -47,7 +47,7 @@ class TestCompetitionEpoch(unittest.TestCase):
 
         
 class MockCompetition(Competition):
-    def __init__(self, override_weights: Union[Decimal, None] = None):
+    def __init__(self, override_weights: Union[float, None] = None):
         self.override_weights = override_weights
     
     def compete(self, challenger: Network, challenged: Network):
